CryptoPkg: Simplify DSC using module scoped Structured PCDs
Simplify CryptoPkg DSC file by setting Structured PCD
field values for different flavors of the crypto modules
in the scope of each module and build all flavors in a
single invocation of the build command.  The CRYPTO_SERVICES
define is removed along with the need to run the build
multiple times with different CRYPTO_SERVICES values to
produce all the flavors.

Cc: Jiewen Yao <jiewen.yao@intel.com>
Cc: Jian J Wang <jian.j.wang@intel.com>
Cc: Xiaoyu Lu <xiaoyux.lu@intel.com>
Cc: Guomin Jiang <guomin.jiang@intel.com>
Cc: Bob Feng <bob.c.feng@intel.com>
Cc: Matthew Carlson <macarl@microsoft.com>
Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>