BaseTools/build: Add EDK_SOURCE_DATE_EPOCH
Always define EDK_SOURCE_DATE_EPOCH to the number
of seconds at the time a build is started so it is
available in DSC and INF files as $(EDK_SOURCE_DATE_EPOCH).

If reproducibuild builds are required then this define
should not be used or it should be defined in a DSC file
to a fixed value.

Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>